+++ hacspec-lib/src/math_util/ct_util.rs
@@ -1,0 +1,85 @@
+use crate::prelude::*;
+
+/// Conditional, constant-time swapping.
+/// Returns `(x, y)` if `c == 0` and `(y, x)` if `c == 1`.
+#[inline]
+#[cfg_attr(feature = "use_attributes", internal(hacspec))]
+pub fn cswap_bit<T: Integer + Copy>(x: T, y: T, c: T) -> (T, T) {
+    cswap(x, y, T::default().wrap_sub(c))
+}
+
+/// Conditional, constant-time swapping.
+/// Returns `(x, y)` if `c == 0` and `(y, x)` if `c == T::max`.
+/// The return value is undefined if `c` has any other value.
+#[inline]
+#[cfg_attr(feature = "use_attributes", internal(hacspec))]
+pub fn cswap<T: Integer + Copy>(x: T, y: T, c: T) -> (T, T) {
+    let mask = c & (x ^ y);
+    (x ^ mask, y ^ mask)
+}
+
+/// Set bit at position `i` in `x` to `b` if `c` is all 1 and return the restult.
+/// Returns `x` if `c` is `0`.
+#[inline]
+#[cfg_attr(feature = "use_attributes", internal(hacspec))]
+pub fn cset_bit<T: Integer + Copy>(x: T, b: T, i: usize, c: T) -> T {
+    let set = x.set_bit(b, i);
+    let (out, _) = cswap(x, set, c);
+    out
+}
+
+/// Add two numerics if condition `c` is set (all bits 1).
+/// Returns `x` if condition `c` is `0`.
+/// Note: Addition is always wrapping.
+#[inline]
+#[cfg_attr(feature = "use_attributes", internal(hacspec))]
+pub fn cadd<T: Integer + Copy>(x: T, y: T, c: T) -> T {
+    let sum = x.wrap_add(y);
+    let (x, _) = cswap(x, sum, c);
+    x
+}
+
+/// Subtract two numerics if condition `c` is set (all bits 1).
+/// Returns `x` if condition `c` is `0`.
+/// Note: Addition is always wrapping.
+#[inline]
+#[cfg_attr(feature = "use_attributes", internal(hacspec))]
+pub fn csub<T: Integer + Copy>(x: T, y: T, c: T) -> T {
+    let diff = x.wrap_sub(y);
+    let (x, _) = cswap(x, diff, c);
+    x
+}
+
+/// Multiply two numerics if condition `c` is set (all bits 1).
+/// Returns `x` if condition `c` is `0`.
+/// Note: Multiplication is always wrapping.
+#[inline]
+#[cfg_attr(feature = "use_attributes", internal(hacspec))]
+pub fn cmul<T: Integer + Copy>(x: T, y: T, c: T) -> T {
+    let prod = x.wrap_mul(y);
+    let (x, _) = cswap(x, prod, c);
+    x
+}
+
+/// Constant time division for Numerics.
+/// Note that this function is only constant time if `T` is a secret integer and
+/// hence provides constant time implementations for the used functions.
+#[inline]
+#[cfg_attr(feature = "use_attributes", internal(hacspec))]
+pub fn ct_div<T: Integer + Copy>(a: T, d: T) -> (T, T) {
+    let mut q = T::default();
+    let mut r = T::default();
+    for i in (0..T::NUM_BITS).rev() {
+        r = r << 1;
+        r = r.set(0, a, i);
+        // The code below is equivalent to the following.
+        // if r.greater_than_or_qual(d) {
+        //     r = r - d;
+        //     q = q.set_bit(T::ONE(), i);
+        // }
+        let geq = r.greater_than_or_equal_bm(d);
+        r = csub(r, d, geq);
+        q = cset_bit(q, T::ONE(), i, geq);
+    }
+    (q, r)
+}
