+++ hacspec-lib/src/prelude.rs
@@ -8,7 +8,7 @@ pub use crate::array::*;
 pub use crate::bigint_integers::*;
 pub use crate::machine_integers::*;
 pub use crate::math_integers::*;
-pub use crate::math_util::*;
+pub use crate::math_util::{ct_util::*, *};
 pub use crate::seq::*;
 pub use crate::traits::*;
 pub use crate::transmute::*;
@@ -16,7 +16,7 @@ pub use crate::util::*;
 pub use crate::vec_integers::*;
 pub use crate::vec_integers_public::*;
 pub use crate::vec_integers_secret::*;
-pub use crate::vec_util::*;
+pub(crate) use crate::vec_util::*;
 pub use crate::*;
 
 pub use abstract_integers::*;
@@ -24,7 +24,7 @@ pub use abstract_integers::*;
 pub use hacspec_attributes::*;
 pub use secret_integers::*;
 
-pub use num::{self, BigUint, CheckedSub, Num, Zero, traits::sign::Signed};
+pub use num::{self, traits::sign::Signed, BigUint, CheckedSub, Num, Zero};
 pub use std::num::ParseIntError;
 pub use std::ops::*;
 pub use std::str::FromStr;
